Re-organizes code for continued refactoring
